alteracoes pasta django_portfolio

+++ django_portfolio/notes/views.py
@@ -49,6 +49,8 @@ def add_list(request):
 @login_required
 def edit_list(request, pk):
     lists_data = get_object_or_404(NoteLists, pk=pk)
+    if lists_data.owner.id!=request.user.id:
+        return redirect('notes:lists')
     form = ListForm(instance=lists_data)
     if request.method=='POST':
         form = ListForm(request.POST, instance=lists_data)
@@ -61,7 +63,8 @@ def edit_list(request, pk):
 @login_required
 def delete_list(request, pk):
     lists_data = get_object_or_404(NoteLists, pk=pk)
-    lists_data.delete()
+    if lists_data.owner.id==request.user.id:
+        lists_data.delete()
     return redirect('notes:lists')
 
 # @login_required
